simple implementation fix
