--- Database.json
@@ -1,1 +1,0 @@
-{"currentStage":{"stage":-1,"minRemaining":30},"activeDMs":{}}
