+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "PomodoroBot",
+  "name": "bot-pomodoro",
   "version": "1.0.0",
   "description": "",
   "main": "botpomodoro.js",
